Rename session module to bluetooth.
